fix types and move utils
